chore(site): refactor dormant badge and add stories
